Moved functionality to blixt_utils, and project specific files to blixt_projects
