+++ unit_tests/test_gassmann.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import matplotlib.pyplot as plt
 
-from rp_utils.attribdict import AttribDict
+from blixt_utils.misc.attribdict import AttribDict
 import rp.rp_core as rp
 import rp.moduli as moduli
 import core.well as cw
